uncomment non-deterministic unit test
this is a TODO for later consideration: buffering in pipes, and
communicating read errors on the consumer back to the writer. I can't
find a reliable unit test with conventional tools right now, pipe
buffering makes it non-obvious for the writer to determine if the reader
correctly consumed all the data.

not extremely high priority, so just marked TODO and ignored the unit
test.